--- src/main/java/com/camping/site/mapper/exampleMapper.java
@@ -1,8 +1,0 @@
-package com.camping.site.mapper;
-
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface exampleMapper {
-
-}
