+++ src/main/java/com/example/camping/CampingSiteApplication.java
@@ -1,4 +1,4 @@
-package com.camping.site;
+package com.example.camping;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -10,5 +10,4 @@ public class CampingSiteApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(CampingSiteApplication.class, args);
 	}
-
 }
