+++ src/main/java/com/example/camping/controller/FaqController.java
@@ -1,15 +1,15 @@
-package com.camping.site.controller;
+package com.example.camping.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/reservation")
-public class ReservationController {
+@RequestMapping("/faq")
+public class FaqController {
 
     @GetMapping
-    public String reservation() {
-        return "pages/reservation";
+    public String faq() {
+        return "pages/faq";
     }
 }
