+++ src/main/java/com/example/camping/controller/HomeController.java
@@ -1,4 +1,4 @@
-package com.camping.site.controller;
+package com.example.camping.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
