+++ src/main/java/com/example/camping/controller/ReservationController.java
@@ -1,15 +1,15 @@
-package com.camping.site.controller;
+package com.example.camping.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/shop")
-public class ShopController {
+@RequestMapping("/reservation")
+public class ReservationController {
 
     @GetMapping
-    public String shop() {
-        return "pages/shop";
+    public String reservation() {
+        return "pages/reservation";
     }
 }
