+++ src/main/java/com/example/camping/controller/ShopController.java
@@ -1,15 +1,15 @@
-package com.camping.site.controller;
+package com.example.camping.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/faq")
-public class FaqController {
+@RequestMapping("/shop")
+public class ShopController {
 
     @GetMapping
-    public String faq() {
-        return "pages/faq";
+    public String shop() {
+        return "pages/shop";
     }
 }
