+++ src/main/java/com/example/camping/repository/UserMapper.java
@@ -1,0 +1,7 @@
+package com.example.camping.repository;
+
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface UserMapper {
+}
