+++ src/test/java/com/example/camping/CampingSiteApplicationTests.java
@@ -1,4 +1,4 @@
-package com.camping.site;
+package com.example.camping;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
